refactor: get out store object from reducer
